requirements.txt
cupy-cuda11x>=12.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pytest>=7.4.0
streamlit>=1.28.0
pillow>=10.0.0
tqdm>=4.65.0

